Fix Homebrew sync script (#83)
This repo uses different names for the binary and for the brew formula,
unlike our other repos. Introduced a new optional parameter to the
homebrew sync script to allow specifying a formula name. If not
supplied, it will default to the binary name, keeping the same behavior
as before.

Some work remains to get this on par with instance-selector (such as
adding support for Mac M1 builds).